requirements.txt
numpy==1.20.1
matplotlib==3.3.4
pandas==1.2.4
seaborn==0.11.1
scipy==1.6.2
tensorflow==2.3.0
torch==1.8.1
torchvision==0.9.1
pillow==8.2.0
opencv=4.0.1
imageio==2.9.0
nltk==3.6.1
apyori==1.1.2
